fix: change  amount to decimal on all code related to Wallet creation

+++ src/PaymentsService.Api/Program.cs
@@ -89,7 +89,7 @@ app.MapGet("/v1/wallets/{id:int}/movements", async (int id, IMovementRepository 
 app.Run();
 
 public record TokenRequest(string Username);
-public record CreateWalletRequest(string DocumentId, string Name, double InitialBalance);
+public record CreateWalletRequest(string DocumentId, string Name, decimal InitialBalance);
 public record TransferRequest(int FromWalletId, int ToWalletId, decimal Amount);
 
 public partial class Program;
